Scene_3
Ground pound practice icicle, destrucable creates added.
