--- Assets/ForToday.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ff8ae0d95e046d94eabfa2944e3908d8
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
