--- Assets/LevelUpperDec.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4dfe2adf38bb8b9408a3380f8f25a10e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
