--- Assets/Materials/Textures.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 12d47bc4105ce65468520a8e62218915
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
