+++ Assets/Materials/Tutorial.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6816b2c50d31cb546a8b3fb243873e8e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
