--- Assets/Materials/TutorialMaterials/Flat_Rock.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d364a0c9d034e764cb5bcb5fd6a5773f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
