--- Assets/Materials/TutorialMaterials/Lily.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4410cb5804a7b574fb812141918e249d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
