--- Assets/NEW_DIR/Boss/Debug.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 57eee7e61b60acb44b41fe166a332be0
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
