--- Assets/NEW_DIR/Boss/Particle.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 191b606a60750904ab0399560be83861
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
