--- Assets/NEW_DIR/Boss/Tutorial.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 499ebf506e0b1634f99f399fddfa1f21
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
