--- Assets/NEW_DIR/Debug/LouiDebugFolder_Please_Do_Not_Alter/Sound.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 315860f1e3c8fcf47a8d30b64b1defb2
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
