--- Assets/NEW_DIR/Enemy/Anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6c67a94c2b56d0b49851aa181f326863
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
