--- Assets/NEW_DIR/Enemy/PhysMat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4954b3725ea7a4549b4a5cec795b6df5
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
