--- Assets/NEW_DIR/Enemy/Sound.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 849617572735cb04b90be8c8263c68c3
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
