--- Assets/NEW_DIR/Enemy/Tex.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 962912b78d616784fad4eeca11bda75b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
