--- Assets/NEW_DIR/Environment/Debug.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f5f1d4a0b082e6a4288135073b40e8a5
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
