--- Assets/NEW_DIR/Environment/Particle.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a9dda4b70fe9c6d4ba78c965bba4b894
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
