--- Assets/NEW_DIR/Environment/PhysMat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e57e0428407456144b71dd4c3994dfa8
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
