--- Assets/NEW_DIR/Environment/Sound.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a8a8c50cda157e446ad4e0f367444798
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
