--- Assets/NEW_DIR/Environment/Tex/Big_Flower_1.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 95f4951432b87d543867a76376754fd1
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
