--- Assets/NEW_DIR/Player/Debug.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8f86fed656a486d4fafc0ee7b716e30b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
