--- Assets/NEW_DIR/Player/PhysMat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f4ce1e83f4895bf44be993a9a359e1ed
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
