--- Assets/NEW_DIR/Player/Sound.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a7a18abdb0b4e384abba410fce48f3af
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
