--- Assets/NEW_DIR/Scenes/Levels_In_Build.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0db50a8b9df1c7444bacc32f0f8e1293
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
