--- Assets/NEW_DIR/Scenes/MapImport.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 770cbf9b15bfdfa42ba52bc58dde17ea
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
