--- Assets/NEW_DIR/Shader/Debug/MatTemplate.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c6890c6dd5a90464a83cc222de81623a
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
