--- Assets/NEW_DIR/Shader/Script.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 64ba3997a3ea873439a4b621f5ea7c7c
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
