--- Assets/NEW_DIR/UI/Debug.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 50fa3d61014e90b49999c4d6386ec7e8
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
