--- Assets/NEW_DIR/UI/Particle.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0d5327ae147fea548989ed947a0eb38c
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
