--- Assets/NEW_DIR/UI/PhysMat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a1f19d7ceb86cd54fa68c7b5b7d4f52d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
