--- Assets/NEW_DIR/UI/Tex.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: fcf79f71931b9cf4a9d5799643070784
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
