+++ Assets/Snow/Textures.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5af0203f45478e94094f805caf69635b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
