--- Assets/_A_Puzzle_Finals/Animations/Icicles.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 37eef42d9dd25c14e9b5997156b55e2c
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
