--- Assets/_A_Puzzle_Finals/Decorations/Trees.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2fe5452dcbb44b24dac444b35e67d606
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
