--- Assets/_A_Puzzle_Finals/Helperes.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9066492783c86274cbcb6de22035b79e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
