--- Assets/_A_Puzzle_Finals/S3_Old.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f2a626b9008eb5e4191f556b4b3f624a
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
