--- Assets/_A_Puzzle_Finals/S3_Scripts/IciclePlatforms.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c4676a7f8ae9995468b913265516c899
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
