+++ Assets/_A_Puzzle_Finals/S3_Scripts/PracticeIcicles.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3ec6a60f28799b149a088d0c1a7d57e4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
