--- Assets/_A_Puzzle_Finals/Scripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d509cc9ff90ea504db96eb5b29ae5cfb
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
